some warnings removed

+++ trunk/utility/src/stream.cpp
@@ -60,7 +60,7 @@ IOStream::~IOStream(){
 bool IStream::readAll(char *_pd, uint32 _dl, uint32){
 	int rv;
 	char *pd = (char*)_pd;
-	while(_dl && (rv = this->read(pd, _dl)) != _dl){
+	while(_dl && (rv = this->read(pd, _dl)) != (int)_dl){
 		if(rv > 0){
 			pd += rv;
 			_dl -= rv;
